requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.12.0
python-dotenv>=1.0.0
ollama>=0.1.7
pyyaml>=6.0
